+++ docs/examples/e06_simple_bot_structure/main.py
@@ -4,7 +4,10 @@ from hata.ext.extension_loader import EXTENSION_LOADER
 
 config = dotenv_values('.env')
 
-Sakuya = Client(config['TOKEN'])
+Sakuya = Client(config['TOKEN'],
+    extensions = 'commands_v2',
+    prefix = '!',
+)
 
 @Sakuya.events
 async def ready(client):